Finished Week 1

+++ bomberman/XYCoordinates.cs
@@ -8,13 +8,13 @@ namespace bomberman
 {
     class XYCoordinates
     {
-        public XYCoordinates(int x, int y)
+        public XYCoordinates(double x, double y)
         {
             X = x;
             Y = y;
         }
 
-        public int X { get; }
-        public int Y { get; }
+        public double X { get; }
+        public double Y { get; }
     }
 }
